+++ cstar/base/utils.py
@@ -1,7 +1,7 @@
-import re
+import functools
 import hashlib
-import warnings
 import subprocess
+from os import PathLike
 from pathlib import Path
 
 
@@ -48,150 +48,6 @@ def _update_user_dotenv(env_file_str) -> None:
 
     with open(Path("~/.cstar.env").expanduser(), "a") as F:
         F.write(env_file_str)
-
-
-def _clone_and_checkout(
-    source_repo: str, local_path: str | Path, checkout_target: str
-) -> None:
-    """Clone `source_repo` to `local_path` and checkout `checkout_target`."""
-    clone_result = subprocess.run(
-        f"git clone {source_repo} {local_path}",
-        shell=True,
-        capture_output=True,
-        text=True,
-    )
-    if clone_result.returncode != 0:
-        raise RuntimeError(
-            f"Error {clone_result.returncode} when cloning repository "
-            + f"{source_repo} to {local_path}. Error messages: "
-            + f"\n{clone_result.stderr}"
-        )
-    print(f"Cloned repository {source_repo} to {local_path}")
-
-    checkout_result = subprocess.run(
-        f"git checkout {checkout_target}",
-        cwd=local_path,
-        shell=True,
-        capture_output=True,
-        text=True,
-    )
-    if checkout_result.returncode != 0:
-        raise RuntimeError(
-            f"Error {checkout_result.returncode} when checking out "
-            + f"{checkout_target} in git repository {local_path}. Error messages: "
-            + f"\n{checkout_result.stderr}"
-        )
-    print(f"Checked out {checkout_target} in git repository {local_path}")
-
-
-def _get_repo_remote(local_root: str | Path) -> str:
-    """Take a local repository path string (local_root) and return as a string the
-    remote URL."""
-    return subprocess.run(
-        f"git -C {local_root} remote get-url origin",
-        shell=True,
-        capture_output=True,
-        text=True,
-    ).stdout.strip()
-
-
-def _get_repo_head_hash(local_root: str | Path) -> str:
-    """Take a local repository path string (local_root) and return as a string the
-    commit hash of HEAD."""
-    return subprocess.run(
-        f"git -C {local_root} rev-parse HEAD",
-        shell=True,
-        capture_output=True,
-        text=True,
-    ).stdout.strip()
-
-
-def _get_hash_from_checkout_target(repo_url: str, checkout_target: str) -> str:
-    """Take a git checkout target (any `arg` accepted by `git checkout arg`) and return
-    a commit hash.
-
-    This method parses the output of `git ls-remote {repo_url}` to create a dictionary
-    of refs and hashes, returning the hash corresponding to `checkout_target` or
-    raising an error listing available branches and tags if the target is not found.
-
-    Parameters:
-    -----------
-    repo_url: str
-        URL pointing to a git-controlled repository
-    checkout_target: str
-        Any valid argument that can be supplied to `git checkout`
-
-    Returns:
-    --------
-    git_hash: str
-        A git commit hash associated with the checkout target
-    """
-
-    # Get list of targets from git ls-remote
-    ls_remote = subprocess.run(
-        f"git ls-remote {repo_url}",
-        shell=True,
-        capture_output=True,
-        text=True,
-    ).stdout
-
-    # Process the output into a `reference: hash` dictionary
-    ref_dict = {
-        ref: has for has, ref in (line.split() for line in ls_remote.splitlines())
-    }
-
-    # If the checkout target is a valid hash, return it
-    if checkout_target in ref_dict.values():
-        return checkout_target
-
-    # Otherwise, see if it is listed as a branch or tag
-    for ref, has in ref_dict.items():
-        if (
-            ref == f"refs/heads/{checkout_target}"
-            or ref == f"refs/tags/{checkout_target}"
-        ):
-            return has
-
-    # Lastly, if NOTA worked, see if the checkout target is a 7 or 40 digit hexadecimal string
-    is_potential_hash = bool(re.fullmatch(r"^[0-9a-f]{7}$", checkout_target)) or bool(
-        re.fullmatch(r"^[0-9a-f]{40}$", checkout_target)
-    )
-    if is_potential_hash:
-        warnings.warn(
-            f"C-STAR: The checkout target {checkout_target} appears to be a commit hash, "
-            f"but it is not possible to verify that this hash is a valid checkout target of {repo_url}"
-        )
-
-        return checkout_target
-
-    # If the target is still not found, raise an error listing branches and tags
-    branches = [
-        ref.replace("refs/heads/", "")
-        for ref in ref_dict
-        if ref.startswith("refs/heads/")
-    ]
-    tags = [
-        ref.replace("refs/tags/", "")
-        for ref in ref_dict
-        if ref.startswith("refs/tags/")
-    ]
-
-    error_message = (
-        f"Supplied checkout_target ({checkout_target}) does not appear "
-        f"to be a valid reference for this repository ({repo_url}).\n"
-    )
-    if branches:
-        error_message += (
-            "Available branches:\n"
-            + "\n".join(f" - {branch}" for branch in sorted(branches))
-            + "\n"
-        )
-    if tags:
-        error_message += (
-            "Available tags:\n" + "\n".join(f" - {tag}" for tag in sorted(tags)) + "\n"
-        )
-
-    raise ValueError(error_message.strip())
 
 
 def _replace_text_in_file(file_path: str | Path, old_text: str, new_text: str) -> bool:
@@ -338,3 +194,89 @@ def _dict_to_tree(input_dict: dict, prefix: str = "") -> str:
                 tree_str += f"{prefix}{sub_prefix}{item_branch}{item}\n"
 
     return tree_str
+
+
+def _run_cmd(
+    cmd: str,
+    cwd: Path | None = None,
+    env: dict[str, str] | None = None,
+    msg_pre: str | None = None,
+    msg_post: str | None = None,
+    msg_err: str | None = None,
+    raise_on_error: bool = False,
+) -> str:
+    """Execute a subprocess using default configuration, blocking until it completes.
+
+    Parameters:
+    -----------
+    cmd (str):
+       The command to be executed as a separate process.
+    cwd (Path, default = None):
+       The working directory for the command. If None, use current working directory.
+    env (Dict[str, str], default = None):
+       A dictionary of environment variables to be passed to the command.
+    msg_pre (str  | None), default = None):
+       An overridden message logged before the command is executed.
+    msg_post (str | None), default = None):
+        An overridden message logged after the command is successfully executed.
+    msg_err (str | None), default = None):
+        An overridden message logged when a command returns a non-zero code. Logs
+        will automatically append the stderr output of the command.
+    raise_on_error (bool, default = False):
+        If True, raises a RuntimeError if the command returns a non-zero code.
+
+    Returns:
+    -------
+    stdout: str
+       The captured standard output of the process.
+
+    Examples:
+    --------
+    In: _run_cmd("python foo.py", msg_pre="Running script", msg_post="Script completed")
+    Out: Running script
+         Script completed
+
+    In: _run_cmd("python foo.py")
+    Out: Running command: python foo.py
+         Command completed successfully.
+
+    In: _run_cmd("python return_nonzero.py")
+    Out: Running command: python return_nonzero.py
+         Command `python return_nonzero.py` failed. STDERR: <stderror of foo.py>
+    """
+    print(msg_pre or f"Running command: {cmd}")
+    stdout: str = ""
+
+    fn = functools.partial(
+        subprocess.run,
+        cmd,
+        shell=True,
+        text=True,
+        capture_output=True,
+    )
+
+    kwargs: dict[str, str | PathLike | dict[str, str]] = {}
+    if cwd:
+        kwargs["cwd"] = cwd
+    if env:
+        kwargs["env"] = env
+
+    result: subprocess.CompletedProcess[str] = fn(**kwargs)
+    stdout = str(result.stdout).strip() if result.stdout is not None else ""
+
+    if result.returncode != 0:
+        rc_out = f"Return Code: `{result.returncode}`."
+        stderr_out = f"STDERR:\n{result.stderr.strip()}"
+
+        if not msg_err:
+            msg_err = f"Command `{cmd}` failed."
+
+        msg = f"{msg_err} {rc_out} {stderr_out}"
+
+        if raise_on_error:
+            raise RuntimeError(msg)
+
+        print(msg)
+
+    print(msg_post or "Command completed successfully.")
+    return stdout